extends UserRepository and GroupRepository from JpaRepository

+++ src/main/java/com/interlink/quiz/controller/QuestionController.java
@@ -64,7 +64,7 @@ public class QuestionController {
         byte[] decode = Base64.getDecoder().decode(url);
         String test = new String(decode);
         CuratorQuiz curatorQuiz = new ObjectMapper().readValue(test, CuratorQuiz.class);
-        QuizDto quizDto = questionService.getQuestionsToGroup(curatorQuiz, userId, httpSession, curatorQuiz.getGroupId(), url);
+        QuizDto quizDto = questionService.getQuestionsToGroup(curatorQuiz, userId, httpSession, curatorQuiz.getGroupId().intValue(), url);
 
         return new ResponseEntity<>(quizDto, HttpStatus.OK);
     }
